Test core splitted to a proper test modules.
